Converted all JSX to TSX.
