+++ index.html
@@ -15,6 +15,6 @@
   </head>
   <body>
     <div id="root"></div>
-    <script type="module" src="/src/entry-client.jsx"></script>
+    <script type="module" src="/src/entry-client.tsx"></script>
   </body>
 </html>
