+++ package.json
@@ -11,7 +11,7 @@
     "lint:css": "stylelint ./src/**/*.{css,sass,scss}",
     "build": "yarn build:client && yarn build:server && yarn build:worker",
     "build:client": "vite build --outDir dist/client --manifest",
-    "build:server": "vite build --outDir dist/server --ssr src/entry-server.jsx",
+    "build:server": "vite build --outDir dist/server --ssr src/entry-server.tsx",
     "build:worker": "cross-env WORKER=true vite build --outDir dist/worker --ssr worker.js",
     "serve": "node --enable-source-maps server"
   },
