+++ src/components/FeaturedCollection.tsx
