+++ src/components/Footer.server.tsx
