+++ src/components/ProductDetails.client.tsx
