+++ src/pages/collections/[handle].server.tsx
