+++ worker.js
@@ -1,5 +1,5 @@
 import handleEvent from '@shopify/hydrogen/worker';
-import entrypoint from './src/entry-server.jsx';
+import entrypoint from './src/entry-server.tsx';
 // eslint-disable-next-line node/no-missing-import
 import indexHtml from './dist/client/index.html?raw';
